Settings: default scaling to Bilinear; move interpolation mapper to Utils
Change the default ImageScalingQuality from HighQualityCubic to Linear
(Bilinear) in AppSettings and appsettings.json, and relocate
ImageInterpolationMapper from Display/ImageRendering to Infra/Utils.

refer #166
