+++ Src/FlyPhotos/appsettings.json
@@ -11,7 +11,7 @@
     "AutoFade": true,
     "FadeIntensity": 60,
     "OpenExitZoom": false,
-    "ImageScalingQuality": "HighQualityCubic",
+    "ImageScalingQuality": "Linear",
     "CheckeredBackground": false,
     "ImageFitPercentage": 100,
     "StretchSmallImages": false,
